gc: introduce incremental mark-sweep GC, the early stage (#36)
The plan is using (SATB) snapshot at the beginning algorithm, Yuasa write barrier.
The write barrier is not added yet, because currently the only place I can think about is the vector-set! function.
set-cons! is not support in this language.
symbol set is in the root and handled by the snapshot.