runtime: let try stack to be the first of the new segment stack
Conceptually, (try chunk handler), thunk belongs to the try block, while handler does not.
In chunk, we can throw twice, they will be handled by the same try, so chunk belongs to the try.
    (try (lambda ()
             (begin (throw 1) (throw 2)))
         (lambda (v k) (k 42)))
If we throw in handler however, it's panic in panic and cannot be catch by this try.
    (try (lambda () (throw 1))
         (lambda (v k) (throw 2)))