eval,runtime: support calling interpreted (lambda ..) from compiled code (#107)
At the very beginning, after compile to C is support,
I got the idea to implement a interpreter in cora, and compile that interpreter to c, 
then I can get an interpreter almost for free to support REPL.
But later I realized that the performance is not good enough.

> Peformance of inspect.cora is bad, so codegen a interpreter written in cora maybe not the best solution.

Then I turn to another way to support REPL in https://github.com/tiancaiamao/cora/pull/27
`eval0` is introduced and it only support symbol, const, and calling compiled code.

Later I found that without `(lambda ...)`, the REPL is inconvenient
So it's added in https://github.com/tiancaiamao/cora/pull/81/
But it's a flawed implementation with a limitation: compiled code can not invoke `(lambda ...)` expression provided from the interpreter.

This commit solve that problem and now there is no such limitation any more.
It's free to mix interpreted code and compiled code, although the performance for the prior case is poor.