*: refactor for the new runtime and update the code generation (#119)
Before this refactor, the generated code is one huge dispatch and basic blocks from all lambda are put into the dispatch.

Now the generated code will keep the original funcation.
This is what fib funcation looks like:

static void clofun0(struct Cora* co, int label, Obj *R) {
	switch (label) {
	case 0:
		{
			Obj n = R[1];
			Obj x105553180305735 = PRIM_LT(n, MAKE_NUMBER(2));
			if (True == x105553180305735) {
				coraReturn(co, n);
				return;
			} else {
				Obj x105553180306375 = PRIM_SUB(n, MAKE_NUMBER(1));
				R[1] = n;
				struct frame1 __curr = {
					.fn = clofun0,
					.label = 2,
					.bp = R,
					.sp = co->ctx.sp,
				};
				vecAppend(&co->callstack, __curr);
				coraCall1(co, globalRef(__symbolTable[0]), x105553180306375);
				return;
			}
		}
	case 1:
		{
			Obj x105553180307143= co->res;
			Obj x105553180306503 = R[1];
			Obj x105553180307335 = PRIM_ADD(x105553180306503, x105553180307143);
			coraReturn(co, x105553180307335);
			return;
		}
	case 2:
		{
			Obj x105553180306503= co->res;
			Obj n = R[1];
			Obj x105553180307015 = PRIM_SUB(n, MAKE_NUMBER(2));
			R[1] = x105553180306503;
			struct frame1 __curr = {
				.fn = clofun0,
				.label = 1,
				.bp = R,
				.sp = co->ctx.sp,
			};
			vecAppend(&co->callstack, __curr);
			coraCall1(co, globalRef(__symbolTable[0]), x105553180307015);
			return;
		}
	}
}

There is the description of the new runtime https://zhuanlan.zhihu.com/p/1945855429440700507
With -O2 compile, the current performance is better than before.

GC is temporarily disabled in this commit, the following PRs should fix them.
This is the first step for the refactor, cleanup is on the way.