*: make GC per Cora, avoid using C stack GC Root (#124)
In the past, each pthread bind with GC, because GC relies on the C stack.
And the parallel model has to be Cora per pthread because of the GC binding.

Now each Cora instance has its own GC (and memory management).
trampoline is the safepoint for GC mark root, C stack should not be used at that point.
Cora manages all its state without interfering the host language (C for example)

This is giant step towards better parallel model.